requirements.txt
asgiref==3.7.2
attrs==23.1.0
Authlib==1.2.1
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
distlib==0.3.6
Django==5.0
djangorestframework==3.14.0
filelock==3.12.0
Flask==3.0.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
platformdirs==3.5.1
pycparser==2.21
pytz==2022.6
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.23.0
Werkzeug==3.0.1
zeep==4.2.1
